parametermanager/snippets-adk/agent_global/requirements.txt
google-adk[extensions]>=1.30
